At Most 10 Matches In A Day
Added feature that forbids person to chose more than 10 matches in a
day. Also added toast messages after a user selects a match (either the
choice is fine, or the match is already in the list, or it can be
already finished, or he may have reached his daily maximum already).